requirements.txt
opencv-python
Pillow
numpy
manga-ocr
ultralytics
huggingface_hub
pytesseract
langdetect
deep-translator